+++ server/odc-migrate/src/main/resources/migrate/common/V_4_3_2_2__alter_connect_connection.sql
@@ -2,4 +2,4 @@
 -- Add column `catalog_name` to `connect_connection` table
 --
 alter table `connect_connection` add column `catalog_name` varchar(255) DEFAULT NULL COMMENT 'Parameters specific to PostgreSql mode, which represent an instance of a database';
-alter table `connect_database` add column `connect_type` varchar(128) DEFAULT NULL COMMENT 'database connect type';
+alter table `connect_database` add column `connect_type` varchar(128) DEFAULT NULL COMMENT 'database connect type, enum values: OB_MYSQL, OB_ORACLE, CLOUD_OB_MYSQL, CLOUD_OB_ORACLE, etc.';
