+++ server/odc-server/src/main/java/com/oceanbase/odc/server/web/controller/v2/DBMetadataController.java
@@ -51,7 +51,7 @@ public class DBMetadataController {
     public ListResponse<SchemaIdentities> listIdentities(@PathVariable String sessionId,
             @RequestParam(required = false, name = "type") List<DBObjectType> types) {
         ConnectionSession session = sessionService.nullSafeGet(SidUtils.getSessionId(sessionId), true);
-        if (ConnectionSessionUtil.getLogicalSession(session)) {
+        if (ConnectionSessionUtil.isLogicalSession(session)) {
             return Responses.list(ListUtils.EMPTY_LIST);
         }
         return Responses.list(identitiesService.list(session, types));
